--- a/benchmark/CMakeLists.txt
+++ b/benchmark/CMakeLists.txt
@@ -4,6 +4,7 @@ add_executable(
         cm_sketch_benchmark.cpp
         tiny_lfu_cms_benchmark.cpp
         large_page_benchmark.cpp
+        bloom_filter_benchmark.cpp
 )
 
 target_include_directories(