--- a/benchmark/large_page_benchmark.cpp
+++ b/benchmark/large_page_benchmark.cpp
@@ -15,21 +15,6 @@ static void LargePage_StoreLoad(benchmark::State& state) {
 BENCHMARK(LargePage_StoreLoad);
 
 /*
-2025-03-10T13:36:41+03:00
-Running ./build_release/benchmark/cache_benchmark
-Run on (16 X 5068.09 MHz CPU s)
-CPU Caches:
-  L1 Data 32 KiB (x8)
-  L1 Instruction 32 KiB (x8)
-  L2 Unified 1024 KiB (x8)
-  L3 Unified 16384 KiB (x1)
-Load Average: 1.31, 1.06, 1.30
---------------------------------------------------------------
-Benchmark                    Time             CPU   Iterations
---------------------------------------------------------------
-LargePage_StoreLoad    4614507 ns      4613506 ns          151
-
-2025-03-10T13:29:41+03:00
 Running ./build_release/benchmark/cache_benchmark
 Run on (16 X 5065.12 MHz CPU s)
 CPU Caches: