### LargePage
Benchmark: https://github.com/mnink275/effective_stream_cache/blob/ink/benchmark/large_page_benchmark.cpp

```
BEFORE:
Run on (16 X 5068.83 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x8)
  L1 Instruction 32 KiB (x8)
  L2 Unified 1024 KiB (x8)
  L3 Unified 16384 KiB (x1)
Load Average: 3.84, 1.99, 1.50
--------------------------------------------------------------
Benchmark                    Time             CPU   Iterations
--------------------------------------------------------------
LargePage_StoreLoad    4719521 ns      4715877 ns          152

AFTER:
Run on (16 X 4920.93 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x8)
  L1 Instruction 32 KiB (x8)
  L2 Unified 1024 KiB (x8)
  L3 Unified 16384 KiB (x1)
Load Average: 1.25, 1.63, 1.43
--------------------------------------------------------------
Benchmark                    Time             CPU   Iterations
--------------------------------------------------------------
LargePage_StoreLoad     606650 ns       606178 ns         1150
```

### BloomFilter
Benchmark: https://github.com/mnink275/effective_stream_cache/blob/ink/benchmark/bloom_filter_benchmark.cpp

```
BEFORE:
Run on (16 X 5064.85 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x8)
  L1 Instruction 32 KiB (x8)
  L2 Unified 1024 KiB (x8)
  L3 Unified 16384 KiB (x1)
Load Average: 1.46, 1.46, 1.40
-----------------------------------------------------------
Benchmark                 Time             CPU   Iterations
-----------------------------------------------------------
BloomFilter_Add        7.76 ns         7.76 ns     78515846
BloomFilter_Test       1.71 ns         1.71 ns    394182553

AFTER:
Run on (16 X 4949.96 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x8)
  L1 Instruction 32 KiB (x8)
  L2 Unified 1024 KiB (x8)
  L3 Unified 16384 KiB (x1)
Load Average: 2.18, 1.66, 1.47
-----------------------------------------------------------
Benchmark                 Time             CPU   Iterations
-----------------------------------------------------------
BloomFilter_Add        1.75 ns         1.75 ns    392168096
BloomFilter_Test       1.82 ns         1.81 ns    387862054
```